+++ .gitignore
@@ -1,3 +1,3 @@
 venv\
-*.pyc
+.pyc
 .env
